backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
gunicorn>=21.2.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# ML & Embeddings
sentence-transformers>=2.3.0
chromadb>=0.4.0
anthropic>=0.18.0

# Text Processing
langchain>=0.1.0
langchain-community>=0.0.17
tiktoken>=0.6.0

# Async Processing
celery>=5.3.0
redis>=5.0.0
django-celery-results>=2.5.0

# Development & Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-mock>=3.12.0
reportlab>=4.0.0
factory-boy>=3.3.0
